feat: create endpoints

+++ api-hooks/user/access/use-revoke-all-access.ts
@@ -1,14 +1,13 @@
 import {useMutation, useQueryClient} from "@tanstack/react-query";
 import {api} from "@/lib/axios";
+import {ENDPOINTS} from "@/shared/config/api";
 
 export const useRevokeAllAccess = () => {
     const queryClient = useQueryClient();
 
     return useMutation({
         mutationFn: async (targetUserId: string) => {
-            const {data} = await api.delete(
-                `/api/users/access/${targetUserId}`
-            );
+            const {data} = await api.delete(ENDPOINTS.revokeAllRights(targetUserId));
 
             return data;
         },
